Move out receipt into dedicated module
