+++ .gitignore
@@ -2,3 +2,4 @@
 /xtask/target
 /xtask/Cargo.lock
 gurk.log
+Session.vim
